refactor : ImageToVideo dto 구조 변경
