+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/ImageToVideo/dto/ImageToVideoRequest.java
@@ -1,4 +1,4 @@
-package com.techeerpicture.TecheerPicture.ImageToVideo;
+package com.techeerpicture.TecheerPicture.ImageToVideo.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
