+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/ImageToVideo/repository/ImageToVideoRepository.java
@@ -1,5 +1,6 @@
-package com.techeerpicture.TecheerPicture.ImageToVideo;
+package com.techeerpicture.TecheerPicture.ImageToVideo.repository;
 
+import com.techeerpicture.TecheerPicture.ImageToVideo.entity.ImageToVideo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
